requirements.txt
python-telegram-bot[job-queue]==20.4
requests==2.30.0
beautifulsoup4